Add files via upload
Problem: Given the head of a singly linked list, reverse the list, and return the reversed list.

Solution: Iterate through the list while adjusting the next pointers of each node, so that each node points to its previous node. Once the iteration is complete, the head of the list will point to the original tail, which becomes the new head of the reversed list. 

Time Complexity: O(n) 
 - We traverse the entire linked list
Space Complexity: O(1)
 - Need a fixed number of variables